rename mark-request-ready
